+++ changelog.md
@@ -17,6 +17,7 @@
 - new widget lib.supersize expands single blocks to full screen width and resizes plots (if available). See docu page for design -> blocks
 - multimedia.image uses new parameter 'localize' to enable URLs containing credentials. If 'true', image is loaded by php script ./lib/multimedia/camimage.php
 - status.toast uses the widget id as additional class name (pure id like in parameter set). This can be used for css styling.
+- new widget plot.xyplot displays x/y data provided by the backend in list item(s)
 
 ### Other New Features
 - template checker now checks smarthomeNG item properties (valid properties and types)